+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - now only a delete button is shown in the bookmarks list if the bookmark
   can be deleted (for [#57](https://github.com/qownnotes/web-companion/issues/57))
 - fixed a problem where links without name couldn't be opened when the link directly was clicked
+- there now is a new private mode switch, that allows opening bookmarks in
+  a private window (for [#58](https://github.com/qownnotes/web-companion/issues/58))
 
 ## 2024.2.3
 - the missing delete code was added (for [#57](https://github.com/qownnotes/web-companion/issues/57))